+++ project/Dependencies.scala
@@ -2,13 +2,13 @@ import sbt._
 
 object Dependencies {
   val akkaV         = "2.6.15"
-  val akkaHttpV     = "10.2.4"
+  val akkaHttpV     = "10.2.6"
   val jacksonV      = "2.12.4"
   val googleV       = "1.22.0"
   val scalaLoggingV = "3.9.4"
   val scalaTestV    = "3.2.9"
   val circeVersion = "0.14.1"
-  val http4sVersion = "0.23.0-RC1"
+  val http4sVersion = "1.0.0-M24"
   val bouncyCastleVersion = "1.69"
 
   def excludeGuavaJDK5(m: ModuleID): ModuleID = m.exclude("com.google.guava", "guava-jdk5")
@@ -35,7 +35,7 @@ object Dependencies {
   val bouncyCastleProviderExt: ModuleID = "org.bouncycastle" % "bcprov-ext-jdk15on" % bouncyCastleVersion
   val bouncyCastleProvider: ModuleID = "org.bouncycastle" % "bcprov-jdk15on" % bouncyCastleVersion
 
-  val catsEffect: ModuleID = "org.typelevel" %% "cats-effect" % "3.1.1"
+  val catsEffect: ModuleID = "org.typelevel" %% "cats-effect" % "3.2.3"
 
   // metrics-scala transitively pulls in io.dropwizard.metrics:metrics-core
   val metricsScala: ModuleID =      "nl.grons"              %% "metrics4-scala"    % "4.1.19"
@@ -59,17 +59,17 @@ object Dependencies {
   val googleRpc: ModuleID =               "io.grpc" % "grpc-core" % "1.34.0"
 
   val googleRpc2: ModuleID =               "io.grpc" % "grpc-core" % "1.34.0"
-  val googleFirestore: ModuleID = "com.google.cloud" % "google-cloud-firestore" % "2.6.1"
-  val googleStorageNew: ModuleID = "com.google.cloud" % "google-cloud-storage" % "1.117.1"
-  val googleStorageLocal: ModuleID = "com.google.cloud" % "google-cloud-nio" % "0.123.3" % "test"
-  val googlePubsubNew: ModuleID = "com.google.cloud" % "google-cloud-pubsub" % "1.113.5"
+  val googleFirestore: ModuleID = "com.google.cloud" % "google-cloud-firestore" % "2.6.2"
+  val googleStorageNew: ModuleID = "com.google.cloud" % "google-cloud-storage" % "1.118.1"
+  val googleStorageLocal: ModuleID = "com.google.cloud" % "google-cloud-nio" % "0.123.4" % "test"
+  val googlePubsubNew: ModuleID = "com.google.cloud" % "google-cloud-pubsub" % "1.113.6"
   val googleKms: ModuleID = "com.google.cloud" % "google-cloud-kms" % "1.43.0"
   val googleDataproc: ModuleID =    "com.google.cloud" % "google-cloud-dataproc" % "1.5.2"
   val googleComputeNew: ModuleID = "com.google.cloud" % "google-cloud-compute" % "1.3.0-alpha"
   val googleContainer: ModuleID = "com.google.cloud" % "google-cloud-container" % "1.5.0"
   val kubernetesClient: ModuleID = "io.kubernetes" % "client-java" % "12.0.0"
-  val googleBigQueryNew: ModuleID = "com.google.cloud" % "google-cloud-bigquery" % "1.134.1"
-  val google2CloudBilling = "com.google.cloud" % "google-cloud-billing" % "1.2.2"
+  val googleBigQueryNew: ModuleID = "com.google.cloud" % "google-cloud-bigquery" % "1.137.2"
+  val google2CloudBilling = "com.google.cloud" % "google-cloud-billing" % "2.0.0"
   val googleResourceManager =  "com.google.cloud" % "google-cloud-resourcemanager" % "0.118.12-alpha"
   //the below v1 module is a dependency for v2 because it contains the OAuth scopes necessary to created scoped credentials
   val googleContainerV1: ModuleID = "com.google.apis" % "google-api-services-container" % "v1-rev20210617-1.32.1"