+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/mock/FakeGoogleComputeService.scala
@@ -48,8 +48,8 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
     instanceName: InstanceName,
     metadataToAdd: Map[String, String],
     metadataToRemove: Set[String]
-  )(implicit ev: Ask[IO, TraceId]): IO[Unit] =
-    IO.unit
+  )(implicit ev: Ask[IO, TraceId]): IO[Option[Operation]] =
+    IO.pure(None)
 
   override def addFirewallRule(project: GoogleProject, firewall: Firewall)(implicit
     ev: Ask[IO, TraceId]
@@ -68,7 +68,7 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def getMachineType(project: GoogleProject, zone: ZoneName, machineTypeName: MachineTypeName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Option[MachineType]] = IO.pure(Some(MachineType.newBuilder().setMemoryMb(7680).build))
+  ): IO[Option[MachineType]] = IO.pure(Some(MachineType.newBuilder().setMemoryMb(7680).setGuestCpus(4).build))
 
   override def getZones(project: GoogleProject, regionName: RegionName)(implicit
     ev: Ask[IO, TraceId]