+++ errorReporting/CHANGELOG.md
@@ -7,4 +7,7 @@ This file documents changes to the `workbench-error-reporting` library, includin
 ### Added
 - Add `ErrorReporting`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-error-reporting" % "0.1-ca18699"`
+### Changed
+- Target java 11
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-error-reporting" % "0.1-TRAVIS-REPLACE-ME"`