+++ errorReporting/CHANGELOG.md
@@ -6,7 +6,12 @@ This file documents changes to the `workbench-error-reporting` library, includin
 Breaking Changes:
 - Upgrade cats-effect to `3.2.3`(see [migration guide](https://typelevel.org/cats-effect/docs/migration-guide#run-the-scalafix-migration)) and a few other dependencies
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-error-reporting" % "0.2-6140478"`
+Dependency Upgrades:
+| Dependency   |      Old Version      |  New Version |
+|----------|:-------------:|------:|
+| google-cloud-errorreporting |  0.120.42-beta | 0.122.9-beta |
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-error-reporting" % "0.2-21c84d8"`
 
 ## 0.1